Move prod tests to its own folder
