+++ test/prod/utils/detectNetwork.js
